refactor(added-courses): split into several comp
