+++ apps/antalmanac/src/components/RightPane/AddedCourses/AddedCoursePane.tsx
@@ -1,385 +1,12 @@
-import { updateScheduleNote } from '$actions/AppStoreActions';
-import { SelectSchedulePopover } from '$components/Calendar/Toolbar/ScheduleSelect/ScheduleSelect';
-import { ClearScheduleButton } from '$components/buttons/Clear';
-import { CopyScheduleButton } from '$components/buttons/Copy';
-import { EmptyState } from '$components/EmptyState';
-import { CustomEventDetailView } from '$components/RightPane/AddedCourses/CustomEventDetailView';
-import { getMissingSections } from '$components/RightPane/AddedCourses/getMissingSections';
-import { NotificationsDialog } from '$components/RightPane/AddedCourses/Notifications/NotificationsDialog';
-import { ColumnToggleDropdown } from '$components/RightPane/CoursePane/CoursePaneButtonRow';
-import SectionTable from '$components/RightPane/SectionTable/SectionTable';
-import { useIsMobile } from '$hooks/useIsMobile';
+import { AddedSectionsGrid } from '$components/RightPane/AddedCourses/AddedSectionsGrid';
+import { FallbackSchedule } from '$components/RightPane/AddedCourses/FallbackSchedule';
 import analyticsEnum, { logAnalytics } from '$lib/analytics/analytics';
-import { clickToCopy } from '$lib/helpers';
-import AppStore from '$stores/AppStore';
 import { useFallbackStore } from '$stores/FallbackStore';
-import { scheduleComponentsToggleStore } from '$stores/ScheduleComponentsToggleStore';
-import { useTabStore } from '$stores/TabStore';
-import { MenuBook } from '@mui/icons-material';
-import { Box, Chip, Paper, SxProps, TextField, Tooltip, Typography } from '@mui/material';
-import { AACourse, SCHEDULE_NOTE_MAX_LENGTH } from '@packages/antalmanac-types';
+import { Box } from '@mui/material';
 import { usePostHog } from 'posthog-js/react';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect } from 'react';
 
-/**
- * All the interactive buttons have the same styles.
- */
-const buttonSx: SxProps = {
-    backgroundColor: 'rgba(236, 236, 236, 1)',
-    marginRight: 1,
-    padding: 1.5,
-    boxShadow: '2',
-    color: 'black',
-    '&:hover': {
-        backgroundColor: 'grey',
-    },
-    pointerEvents: 'auto',
-};
-
-export interface CourseWithTerm extends AACourse {
-    term: string;
-}
-
-function getCourses() {
-    const currentCourses = AppStore.schedule.getCurrentCourses();
-
-    const formattedCourses: CourseWithTerm[] = [];
-
-    for (const course of currentCourses) {
-        let formattedCourse = formattedCourses.find(
-            (needleCourse) =>
-                needleCourse.courseNumber === course.courseNumber &&
-                needleCourse.deptCode === course.deptCode &&
-                needleCourse.courseTitle === course.courseTitle
-        );
-
-        const sectionUpdatedAt = course.section?.updatedAt ?? null;
-
-        if (formattedCourse) {
-            formattedCourse.sections.push({
-                ...course.section,
-            });
-            formattedCourse.updatedAt = sectionUpdatedAt;
-        } else {
-            formattedCourse = {
-                term: course.term,
-                deptCode: course.deptCode,
-                courseComment: course.courseComment,
-                prerequisiteLink: course.prerequisiteLink,
-                courseNumber: course.courseNumber,
-                courseTitle: course.courseTitle,
-                sectionTypes: course.sectionTypes,
-                sections: [
-                    {
-                        ...course.section,
-                    },
-                ],
-                updatedAt: sectionUpdatedAt ?? null,
-            };
-            formattedCourses.push(formattedCourse);
-        }
-    }
-
-    formattedCourses.forEach(function (course) {
-        course.sections.sort(function (a, b) {
-            return parseInt(a.sectionCode, 10) - parseInt(b.sectionCode, 10);
-        });
-    });
-
-    return formattedCourses;
-}
-
-function CustomEventsBox() {
-    const { fallbackMode, getCurrentFallbackSchedule } = useFallbackStore();
-    const currentScheduleIndex = AppStore.getCurrentScheduleIndex();
-
-    const [customEvents, setCustomEvents] = useState(
-        fallbackMode
-            ? getCurrentFallbackSchedule(currentScheduleIndex).customEvents
-            : AppStore.schedule.getCurrentCustomEvents()
-    );
-
-    useEffect(() => {
-        const handleCustomEventsChange = () => {
-            setCustomEvents([...AppStore.schedule.getCurrentCustomEvents()]);
-        };
-
-        AppStore.on('customEventsChange', handleCustomEventsChange);
-        AppStore.on('currentScheduleIndexChange', handleCustomEventsChange);
-
-        return () => {
-            AppStore.off('customEventsChange', handleCustomEventsChange);
-            AppStore.off('currentScheduleIndexChange', handleCustomEventsChange);
-        };
-    }, []);
-
-    if (customEvents.length <= 0) {
-        return null;
-    }
-
-    return (
-        <Box>
-            <Typography variant="h6">Custom Events</Typography>
-            <Box display="flex" flexDirection="column" gap={1}>
-                {customEvents.map((customEvent) => {
-                    return (
-                        <Box key={customEvent.customEventID}>
-                            <CustomEventDetailView
-                                customEvent={customEvent}
-                                scheduleNames={AppStore.getScheduleNames()}
-                            />
-                        </Box>
-                    );
-                })}
-            </Box>
-        </Box>
-    );
-}
-
-function ScheduleNoteBox() {
-    const { fallbackMode, getCurrentFallbackSchedule } = useFallbackStore();
-    const [scheduleNote, setScheduleNote] = useState(
-        fallbackMode
-            ? getCurrentFallbackSchedule(AppStore.getCurrentScheduleIndex()).scheduleNote
-            : AppStore.getCurrentScheduleNote()
-    );
-    const [scheduleIndex, setScheduleIndex] = useState(AppStore.getCurrentScheduleIndex());
-
-    const handleNoteChange = useCallback(
-        (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-            setScheduleNote(event.target.value);
-            updateScheduleNote(event.target.value, scheduleIndex);
-        },
-        [scheduleIndex]
-    );
-
-    useEffect(() => {
-        const handleScheduleNoteChange = () => {
-            setScheduleNote(AppStore.getCurrentScheduleNote());
-        };
-
-        const handleScheduleIndexChange = () => {
-            setScheduleIndex(AppStore.getCurrentScheduleIndex());
-        };
-
-        AppStore.on('scheduleNotesChange', handleScheduleNoteChange);
-        AppStore.on('currentScheduleIndexChange', handleScheduleIndexChange);
-
-        return () => {
-            AppStore.off('scheduleNotesChange', handleScheduleNoteChange);
-            AppStore.off('currentScheduleIndexChange', handleScheduleIndexChange);
-        };
-    }, []);
-
-    return (
-        <Box>
-            <Typography variant="h6">Schedule Notes</Typography>
-
-            <TextField
-                type="text"
-                color="secondary"
-                variant="filled"
-                label="Click here to start typing!"
-                onChange={handleNoteChange}
-                value={scheduleNote}
-                inputProps={{
-                    maxLength: SCHEDULE_NOTE_MAX_LENGTH,
-                    style: { cursor: fallbackMode ? 'not-allowed' : 'text' },
-                }}
-                InputLabelProps={{
-                    variant: 'filled',
-                }}
-                InputProps={{ disableUnderline: true }}
-                fullWidth
-                multiline
-                disabled={fallbackMode}
-                sx={{
-                    '& .MuiInputBase-root': {
-                        cursor: fallbackMode ? 'not-allowed' : 'text',
-                    },
-                }}
-            />
-        </Box>
-    );
-}
-
-function FallbackSchedule() {
-    const { getCurrentFallbackSchedule } = useFallbackStore();
-    const [currentScheduleIndex, setCurrentScheduleIndex] = useState(AppStore.getCurrentScheduleIndex());
-    const postHog = usePostHog();
-
-    useEffect(() => {
-        const handleScheduleIndexChange = () => {
-            setCurrentScheduleIndex(AppStore.getCurrentScheduleIndex());
-        };
-
-        AppStore.on('currentScheduleIndexChange', handleScheduleIndexChange);
-
-        return () => {
-            AppStore.off('currentScheduleIndexChange', handleScheduleIndexChange);
-        };
-    }, []);
-
-    const fallbackSchedule = getCurrentFallbackSchedule(currentScheduleIndex);
-
-    const sectionsByTerm: [string, string[]][] = useMemo(() => {
-        const result = fallbackSchedule.courses.reduce(
-            (accumulated, course) => {
-                accumulated[course.term] ??= [];
-                accumulated[course.term].push(course.sectionCode);
-                return accumulated;
-            },
-            {} as Record<string, string[]>
-        );
-
-        return Object.entries(result);
-    }, [fallbackSchedule.courses]);
-
-    return (
-        <Box display="flex" flexDirection="column" gap={1}>
-            <Typography variant="h6">{fallbackSchedule.scheduleName}</Typography>
-            {
-                // Sections organized under terms, in case the schedule contains multiple terms
-                sectionsByTerm.map(([term, sections]) => (
-                    <Box key={term}>
-                        <Typography variant="h6">{term}</Typography>
-                        <Paper key={term} elevation={1}>
-                            {sections.map((section, index) => (
-                                <Tooltip title="Click to copy section code" placement="right" key={index}>
-                                    <Chip
-                                        onClick={(event) => {
-                                            clickToCopy(event, section);
-                                            logAnalytics(postHog, {
-                                                category: analyticsEnum.addedClasses,
-                                                action: analyticsEnum.addedClasses.actions.COPY_COURSE_CODE,
-                                            });
-                                        }}
-                                        label={section}
-                                        size="small"
-                                        style={{ margin: '10px 10px 10px 10px' }}
-                                        key={index}
-                                    />
-                                </Tooltip>
-                            ))}
-                        </Paper>
-                    </Box>
-                ))
-            }
-
-            <CustomEventsBox />
-
-            <ScheduleNoteBox />
-
-            <Typography variant="body1">
-                Anteater API is currently unreachable. This is the information that we can currently retrieve.
-            </Typography>
-        </Box>
-    );
-}
-
-function AddedSectionsGrid() {
-    const [courses, setCourses] = useState(getCourses());
-    const [scheduleNames, setScheduleNames] = useState(AppStore.getScheduleNames());
-    const [scheduleIndex, setScheduleIndex] = useState(AppStore.getCurrentScheduleIndex());
-
-    const isMobile = useIsMobile();
-
-    useEffect(() => {
-        const handleCoursesChange = () => {
-            setCourses(getCourses());
-        };
-
-        const handleScheduleNamesChange = () => {
-            setScheduleNames([...AppStore.getScheduleNames()]);
-        };
-
-        const handleScheduleIndexChange = () => {
-            setScheduleIndex(AppStore.getCurrentScheduleIndex());
-        };
-
-        AppStore.on('addedCoursesChange', handleCoursesChange);
-        AppStore.on('currentScheduleIndexChange', handleCoursesChange);
-        AppStore.on('scheduleNamesChange', handleScheduleNamesChange);
-        AppStore.on('currentScheduleIndexChange', handleScheduleIndexChange);
-
-        return () => {
-            AppStore.off('addedCoursesChange', handleCoursesChange);
-            AppStore.off('currentScheduleIndexChange', handleCoursesChange);
-            AppStore.off('scheduleNamesChange', handleScheduleNamesChange);
-            AppStore.off('currentScheduleIndexChange', handleScheduleIndexChange);
-        };
-    }, []);
-
-    const scheduleUnits = useMemo(() => {
-        let result = 0;
-
-        for (const course of courses) {
-            for (const section of course.sections) {
-                if (!isNaN(Number(section.units))) {
-                    result += Number(section.units);
-                }
-            }
-        }
-
-        return result;
-    }, [courses]);
-
-    const scheduleName = useMemo(() => {
-        return scheduleNames[scheduleIndex];
-    }, [scheduleNames, scheduleIndex]);
-
-    return (
-        <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-            <Box sx={{ display: 'flex', width: 'fit-content', position: 'absolute', zIndex: 2 }}>
-                <CopyScheduleButton index={scheduleIndex} buttonSx={buttonSx} />
-                <ClearScheduleButton buttonSx={buttonSx} analyticsCategory={analyticsEnum.addedClasses} />
-                <ColumnToggleDropdown />
-                <NotificationsDialog buttonSx={buttonSx} />
-            </Box>
-            <Box sx={{ marginTop: 7 }}>
-                <Typography variant="h6">{`${scheduleName} (${scheduleUnits} Units)`}</Typography>
-                {isMobile && <SelectSchedulePopover />}
-                {courses.length === 0 && (
-                    <EmptyState
-                        Icon={MenuBook}
-                        title="No Courses Added Yet"
-                        description="Search for courses and add sections to build your schedule. You can also import from your study list."
-                        primaryAction={{
-                            label: 'Search Courses',
-                            onClick: () => useTabStore.getState().setActiveTab('search'),
-                        }}
-                        secondaryAction={{
-                            label: 'Import Schedule',
-                            onClick: () => scheduleComponentsToggleStore.getState().setOpenImportDialog(true),
-                        }}
-                    />
-                )}
-                <Box display="flex" flexDirection="column" gap={1}>
-                    {courses.map((course) => {
-                        const missingSections = getMissingSections(course);
-
-                        return (
-                            <Box key={course.deptCode + course.courseNumber + course.courseTitle}>
-                                <SectionTable
-                                    courseDetails={course}
-                                    term={course.term}
-                                    allowHighlight={false}
-                                    analyticsCategory={analyticsEnum.addedClasses}
-                                    scheduleNames={scheduleNames}
-                                    missingSections={missingSections}
-                                />
-                            </Box>
-                        );
-                    })}
-                </Box>
-            </Box>
-
-            <CustomEventsBox />
-
-            <ScheduleNoteBox />
-        </Box>
-    );
-}
+export type { CourseWithTerm } from '$components/RightPane/AddedCourses/AddedSectionsGrid';
 
 export function AddedCoursePane() {
     const fallbackMode = useFallbackStore((state) => state.fallbackMode);
